% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ISOedit}
\alias{ISOedit}
\alias{MISOedit}
\alias{MISOpull}
\title{Edit ISO objects.}
\usage{
ISOedit(object, ...)

MISOedit(object, rejection = F)

MISOpull(object, index, data.names = NULL)
}
\description{
Edit ISO objects.
}
