% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ISOfind}
\alias{ISOfind}
\title{Class filter list}
\usage{
text
}
\arguments{
\item{ls()}{text. Checks global environment for current ISO/MISO object names.}
}
\value{
text
}
\description{
Class filter list
}
\examples{
test
}
