% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-class.R
\name{ISOgen}
\alias{ISOgen}
\title{Generate ISO-class object.}
\usage{
ISOgen(x)
}
\arguments{
\item{x}{text}
}
\value{
Generates ISO-class object.
}
\description{
Generate ISO-class object.
}
\examples{
text

}
