% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ISOpull}
\alias{ISOpull}
\alias{MISOpull}
\title{Pull values}
\usage{
text

MISOpull(object, index, data.names = NULL)
}
\arguments{
\item{ISOobject}{text. Pull values}
}
\value{
text
}
\description{
Pull values
}
\examples{
test
}
