% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoclock.R
\name{isosub}
\alias{isosub}
\alias{isoreal}
\title{Wrapper function to calculate isotope residence}
\usage{
isosub(params)

isoreal(doi, dfi, dt, lambda)
}
\description{
Wrapper function to calculate isotope residence
}
