% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{miso_resplot}
\alias{miso_resplot}
\title{MISO plotting.}
\usage{
miso_resplot(
  object,
  y,
  tail.prob = TRUE,
  data.names = NULL,
  x = NULL,
  seperator = NULL,
  xlim = NULL,
  ylim = NULL,
  facet = NULL,
  quantile = F,
  tail.summary = F
)
}
\description{
MISO plotting.
}
