% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-class.R
\name{MISOgen}
\alias{MISOgen}
\title{Generate MISO-class object.}
\usage{
MISOgen(x)
}
\arguments{
\item{x}{text}
}
\value{
Generates MISO-class object.
}
\description{
Generate MISO-class object.
}
\examples{
text

}
