% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoclock.R
\name{REScalc}
\alias{REScalc}
\title{Wrapper function to calculate isotope residence for ISO and MISO objects}
\usage{
REScalc(object, index)
}
\description{
Wrapper function to calculate isotope residence for ISO and MISO objects
}
