% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{resplot}
\alias{resplot}
\title{Method for plotting residence times from ISO objects.}
\usage{
resplot(
  object,
  y = y,
  tail.prob = TRUE,
  tail.summary = FALSE,
  data.names = NULL,
  x = NULL,
  seperator = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\description{
Method for plotting residence times from ISO objects.
}
