% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary,ISO-method}
\alias{summary,ISO-method}
\title{Method for what is printed for ISO object summaries.}
\usage{
\S4method{summary}{ISO}(object)

\S4method{summary}{ISO}(object)
}
\description{
Method for what is printed for ISO object summaries.

Method for what is printed for MISO object summaries.
}
