% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoclock.R
\name{isoclock}
\alias{isoclock}
\title{Calculate residence time for an individual.}
\usage{
isoclock(animal)
}
\arguments{
\item{animal}{text}
}
\value{
Residence time (days)
}
\description{
Calculate residence time for an individual.
}
\details{
Calculates the residence time (days) of an individual (\link[isoclockR]{ISO}).
}
\references{
Klaasen, M., Piersma, T., Korthals, H., Dekinga, A., and Dietz, M.W. 2010. Single-point isotope measurements in blood cells and plasma to estimate
the time since diet switches. Functional Ecology, 24: 794-804.

Shipley, O., Newton, A., Frisk, M., Henkes, G., Walters, H., LaBella, J., Hyatt, M., and Olin, J. In Review. Telemetry validated nitrogen stable isotope clocks
identify ocean-to-estuarine habitat shifts in mobile organisms. Methods in Ecology and Evolution.
}
