% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MISO-class}
\alias{MISO-class}
\alias{MISO}
\title{Wrapping class for multiple isotope objects (MISO).}
\description{
An S4-class object that stores both data and metadata for all isotope objects
of interest (i.e., \link[isoclockR]{ISO}).
}
\details{
words
}
\section{Slots}{

\describe{
\item{\code{metadata}}{Any metadata such as datetime, species ID, etc.}

\item{\code{data}}{Any calculations such as residence time estimation.}

\item{\code{timestamp}}{Timestamp of object generation.}
}}

\author{
Brandyn M. Lucca
}
